CMPLRLLVM-10096: use int64_t type for device_num arguments in libomptarget.
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/205965
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>

+++ openmp/libomptarget/include/omptarget.h
@@ -295,18 +295,18 @@ EXTERN
 void __kmpc_push_target_tripcount(int64_t device_id, uint64_t loop_tripcount);
 #if INTEL_COLLAB
 EXTERN
-bool __tgt_is_device_available(int device_num, void *device_type);
+bool __tgt_is_device_available(int64_t device_num, void *device_type);
 EXTERN
-void *__tgt_create_buffer(int device_num, void *host_ptr);
+void *__tgt_create_buffer(int64_t device_num, void *host_ptr);
 EXTERN
-int __tgt_release_buffer(int device_num, void *tgt_buffer);
+int __tgt_release_buffer(int64_t device_num, void *tgt_buffer);
 
 // Returns implementation defined device name for the given device number,
 // using provided Buffer. Buffer must be able to hold at least BufferMaxSize
 // characters. Returns nullptr, if device name cannot be acquired, otherwise,
 // returns a '\0' terminated C string (pointer to Buffer).
 EXTERN char *__tgt_get_device_name(
-    int device_num, char *buffer, size_t buffer_max_size);
+    int64_t device_num, char *buffer, size_t buffer_max_size);
 
 // Returns implementation defined RTL name corresponding to the given
 // device number, using provided Buffer. Buffer must be able to hold
@@ -314,7 +314,7 @@ EXTERN char *__tgt_get_device_name(
 // Returns nullptr, if RTL name cannot be acquired, otherwise,
 // returns a '\0' terminated C string (pointer to Buffer).
 EXTERN char *__tgt_get_device_rtl_name(
-    int device_num, char *buffer, size_t buffer_max_size);
+    int64_t device_num, char *buffer, size_t buffer_max_size);
 #endif // INTEL_COLLAB
 
 #ifdef __cplusplus
